+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), with added/changed/deprecated/removed/fixed/security.
 
 ## [unreleased]
+## [0.3.0] - 2025-03-16
+
+### Added
+* Use of Python module dictionary_en_gcide for lookup words in english
+* Normalize words not found in dictionary with nltk
+
 ## [0.2.2] - 2023-10-17
 
 ### Changed
@@ -38,7 +44,8 @@ All notable changes to this project will be documented in this file. Based on [K
 ### Added
 - Initial app with words api and flashcard creation
 
-[unreleased]: https://github.com/leafarlins/gokopa/compare/v0.2.2...HEAD
+[unreleased]: https://github.com/leafarlins/gokopa/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/leafarlins/lang/compare/v0.2.2..0.3.0/
 [0.2.2]: https://github.com/leafarlins/lang/compare/v0.2.1..0.2.2/
 [0.2.1]: https://github.com/leafarlins/lang/compare/v0.2.0..0.2.1/
 [0.2.0]: https://github.com/leafarlins/lang/compare/v0.1.2..0.2.0/